Seed demon codex from local data
